Consolidate command bar tests
